requirements.txt
fastapi[standard]
sqlalchemy
aiofiles
aioshutil
aioconsole
python-dotenv
pytest
pytest-asyncio